+++ src/datafactory/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.10.0
++++++
+* Support new features in ADF
+
 0.9.0
 +++++
 * Change publicnetworkaccess to public-network-access word